Calculator cost

+++ Spring-07-SpringBoot-loC-Di-Lab/src/main/java/com/cybertek/interfaces/carpetPrices/Carpet.java
@@ -2,7 +2,9 @@ package com.cybertek.interfaces.carpetPrices;
 
 import com.cybertek.enums.City;
 
+import java.math.BigDecimal;
+
 public interface Carpet {
 
-    double getSqFtPrice(City city);
+    BigDecimal getSqFtPrice(City city);
 }
